+++ src/Command/Loader.psm1
@@ -1,1 +1,2 @@
+Invoke-Expression ([System.IO.Path]::Combine($PSScriptRoot, 'Resolver.ps1'))
 Import-Module ([System.IO.Path]::Combine($PSScriptRoot, 'AstralKeks.Workbench.Command.dll'))